feat: `Commit::short_id()` (#301)
